Add magic tags: AI-assigned keyword tags on index
- MagicTags mapping type with fromField, api(), maxTags (default 5)
- NewProperties::magicTags() and Properties::magicTagsFields()
- DocumentProcessor::populateMagicTags with LLM jsonAnswer and index terms context
- AliveCollection: fetchExistingMagicTags, populateMagicTags toggle
- UsesApis::getLlmApi(); NewJsonSchema::stringArray()
- VoyageEmbeddingsApi: cast batch dims to int for array_pad

Made-with: Cursor

+++ src/AI/APIs/VoyageEmbeddingsApi.php
@@ -69,7 +69,7 @@ class VoyageEmbeddingsApi implements EmbeddingsApi
 
         $texts = array_map(fn ($item) => $item['text'] ?? '', $payload);
 
-        $dimensions = $payload[0]['dims'] ?? 0;
+        $dimensions = (int) ($payload[0]['dims'] ?? 0);
 
         $response = $this->client->post('/v1/embeddings', [
             RequestOptions::JSON => [
